make `take_preprocess_calls` symmetric
